Added integration test (#7)
* Updated tests.

* Added vector and string serialization

* Refs: #3308 and #3309. Added arrays n-dim and serialized size functions.

* Refs: #3118. Intermediate working state.

* #3318. Added '-test' option functionality

* Updated submodule IDL-parser

* Minor changes

* Organization: moved stgs

* Refs #3343 #3339 Added integration tests.

* Refs #3339. Version from the build.gradle. Minor fixes.
