+++ .gitignore
@@ -41,3 +41,7 @@ gradle-app.setting
 # gradle/wrapper/gradle-wrapper.properties
 
 # End of https://www.gitignore.io/api/java,gradle
+
+# Eclipse
+.classpath
+.project
