+++ build.gradle
@@ -15,7 +15,8 @@
 apply plugin: 'java'
 apply plugin: 'eclipse' // Eclipse integration
 
-description = """"""
+version = '1.0.0beta2'
+description = 'Build script with all necessary tasks to compile and test MicroRTPSGen'
 
 repositories {
     mavenCentral()
@@ -23,6 +24,7 @@ repositories {
 
 dependencies {
     compile files('thirdparty/IDL-Parser/build/libs/idlparser-1.0.0.jar')
+    testCompile group: 'junit', name: 'junit', version: '4.+'
 }
 
 task buildIDLParser(type: GradleBuild) {
@@ -30,28 +32,18 @@ task buildIDLParser(type: GradleBuild) {
     tasks = ['clean', 'build']
 }
 
-task copyResources {
-    // Create directory
-    File outputDir = file(new File(buildDir, "resources/main"))
-    outputDir.exists() || outputDir.mkdirs()
-
-    // Create version file
-    Properties pversion = new Properties()
-    project.hasProperty('customversion') ? pversion.setProperty('version', project.customversion.toString()) : pversion.setProperty('version', '0.0.0')
-    File versionFile = new File("${project.buildDir}/resources/main/version")
-    versionFile.createNewFile();
-    pversion.store(versionFile.newWriter(), null)
-}
-
 sourceSets {
     main {
         java {
-            srcDir 'src'
+            srcDir 'src/main/java'
         }
-
         resources {
-            srcDir 'src'
-            include '**/*.stg'
+            srcDir 'src/main/resources'
+        }
+    }
+    test {
+        java {
+            srcDir 'src/test/java'
         }
     }
 }
@@ -59,9 +51,10 @@ sourceSets {
 jar {
     from { configurations.compile.collect { it.isDirectory() ? it : zipTree(it) } }
     baseName = 'micrortpsgen'
+    archiveName = "${jar.baseName}.${jar.extension}"
     destinationDir = file(new File(projectDir, 'share/micrortps'))
     manifest {
-        attributes("Created-By": "eProsima", "Main-Class": "com.eprosima.micrortps.micrortpsgen")
+        attributes("Created-By": "eProsima", "Main-Class": "com.eprosima.micrortps.micrortpsgen", "Specification-Version": version)
     }
 }
 
@@ -71,22 +64,8 @@ compileJava {
     targetCompatibility = 1.6
 }
 
-
-task testCodeGen(type: JavaExec) {
-    File outputDir = file("${project.buildDir}/test")
-    outputDir.exists() || outputDir.mkdirs()
-
-    workingDir './'
-
-    main = '-jar';
-    args = [
-        'share/micrortps/micrortpsgen.jar',
-        '-replace',
-        '-example',
-        '-write-access-profile',
-        'test/IncludeType.idl',
-        'test/TestType.idl',
-        '-d',
-        outputDir,
-    ]
+test {
+    useJUnit()
 }
+
+
