--- src/com/eprosima/micrortps/idl/templates/RTPSSubscriberSource.stg
@@ -1,153 +1,0 @@
-// Copyright 2016 Proyectos y Sistemas de Mantenimiento SL (eProsima).
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-group ProtocolHeader;
-
-main(ctx, definitions) ::= <<
-$fileHeader(ctx=ctx,  file=[ctx.filename, "Subscriber.c"], description=["This file contains the implementation of the publisher functions."])$
-
-#include "$ctx.filename$.h"
-#include <micrortps/client/client.h>
-#include <string.h> //strcmp
-#include <stdlib.h> //atoi
-#include <stdio.h>
-
-#define STREAM_HISTORY  8
-#define BUFFER_SIZE     MR_CONFIG_UDP_TRANSPORT_MTU * STREAM_HISTORY
-
-$definitions; separator="\n"$
-
-const char* topic_xml =
-"<dds><topic>$if(ctx.lastStructure.typeCode.hasKey)$<kind>WITH_KEY</kind>$endif$<name>$ctx.filename$PubSubTopic</name><dataType>$ctx.lastStructure.scopedname$</dataType></topic></dds>";
-const char* subscriber_xml = "<subscriber name=\"MySubscriber\"";
-const char* datareader_xml = "<profiles><subscriber profile_name=\"default_xrce_subscriber_profile\"><topic><kind>$if(ctx.lastStructure.typeCode.hasKey)$WITH_KEY$else$NO_KEY$endif$</kind><name>$ctx.filename$PubSubTopic</name><dataType>$ctx.lastStructure.scopedname$</dataType><historyQos><kind>KEEP_LAST</kind><depth>5</depth></historyQos><durability><kind>TRANSIENT_LOCAL</kind></durability></topic></subscriber></profiles>";
-
-void on_topic(mrSession* session, mrObjectId object_id, uint16_t request_id, mrStreamId stream_id, struct MicroBuffer* mb, void* args)
-{
-    (void) session; (void) object_id; (void) request_id; (void) stream_id; (void) args;
-
-    $ctx.lastStructure.cScopedname$ topic;
-    $ctx.lastStructure.cScopedname$_deserialize_topic(mb, &topic);
-
-    printf("Received topic");
-}
-
-int main(void)
-{
-    // Transport
-    mrUDPTransport transport;
-    if(!mr_init_udp_transport(&transport, "127.0.0.1", 2018))
-    {
-        printf("Error at create transport.\n");
-        return 1;
-    }
-
-    // Session
-    mrSession session;
-    mr_init_session(&session, &transport.comm, 0xCCCCDDDD);
-    mr_set_topic_callback(&session, on_topic, NULL);
-    if(!mr_create_session(&session))
-    {
-        printf("Error at create session.\n");
-        return 1;
-    }
-
-    // Streams
-    uint8_t output_reliable_stream_buffer[BUFFER_SIZE];
-    mrStreamId reliable_out = mr_create_output_reliable_stream(&session, output_reliable_stream_buffer, BUFFER_SIZE, STREAM_HISTORY);
-
-    uint8_t input_reliable_stream_buffer[BUFFER_SIZE];
-    mrStreamId reliable_in = mr_create_input_reliable_stream(&session, input_reliable_stream_buffer, BUFFER_SIZE, STREAM_HISTORY);
-
-    // Create entities
-    mrObjectId participant_id = mr_object_id(0x01, MR_PARTICIPANT_ID);
-    uint16_t participant_req = mr_write_create_participant_ref(&session, reliable_out, participant_id, 0, "default participant", MR_REPLACE);
-
-    mrObjectId topic_id = mr_object_id(0x01, MR_TOPIC_ID);
-    uint16_t topic_req = mr_write_configure_topic_xml(&session, reliable_out, topic_id, participant_id, topic_xml, MR_REPLACE);
-
-    mrObjectId subscriber_id = mr_object_id(0x01, MR_SUBSCRIBER_ID);
-    uint16_t subscriber_req = mr_write_configure_subscriber_xml(&session, reliable_out, subscriber_id, participant_id, subscriber_xml, MR_REPLACE);
-
-    mrObjectId datareader_id = mr_object_id(0x01, MR_DATAREADER_ID);
-    uint16_t datareader_req = mr_write_configure_datareader_xml(&session, reliable_out, datareader_id, subscriber_id, datareader_xml, MR_REPLACE);
-
-    // Send create entities message and wait its status
-    uint8_t status[4];
-    uint16_t requests[4] = {participant_req, topic_req, subscriber_req, datareader_req};
-    if(!mr_run_session_until_status(&session, 1000, requests, status, 4))
-    {
-        printf("Error at create entities: participant: %i topic: %i subscriber: %i datareader: %i\n", status[0], status[1], status[2], status[3]);
-        return 1;
-    }
-
-    // Request topics
-    mrDeliveryControl delivery_control = {0};
-    delivery_control.max_samples = MR_MAX_SAMPLES_UNLIMITED;
-    uint16_t read_data_req = mr_write_request_data(&session, reliable_out, datareader_id, reliable_in, &delivery_control);
-
-    // Read topics
-    while(true)
-    {
-        uint8_t read_data_status;
-        (void) mr_run_session_until_status(&session, MR_TIMEOUT_INF, &read_data_req, &read_data_status, 1);
-    }
-
-    // Delete resources
-    mr_delete_session(&session);
-
-    return 0;
-}
-
-
->>
-
-module(ctx, parent, module, definition_list) ::= <<
-$definition_list$
->>
-
-definition_list(definitions) ::= <<
-$definitions; separator="\n"$
-
->>
-
-annotation(ctx, annotation) ::= <<>>
-
-interface(ctx, parent, interface, export_list) ::= <<
-$export_list$
->>
-
-export_list(exports) ::= <<
-$exports; separator="\n"$
-
->>
-
-exception(ctx, parent, exception) ::= <<>>
-
-operation(ctx, parent, operation, param_list) ::= <<>>
-
-param_list(parameters) ::= <<>>
-
-param(parameter) ::= <<>>
-
-const_decl(ctx, parent, const) ::= <<>>
-
-typedef_decl(ctx, parent, typedefs) ::= <<>>
-
-struct_type(ctx, parent, struct) ::= <<
->>
-
-union_type(ctx, parent, union) ::= <<>>
-
-enum_type(ctx, parent, enum) ::= <<>>
