--- src/com/eprosima/micrortps/idl/templates/eprosima.stg
@@ -1,193 +1,0 @@
-// Copyright 2016 Proyectos y Sistemas de Mantenimiento SL (eProsima).
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-group eprosima;
-
-paramTypeByValue(typecode) ::= <<
-$if(typecode.primitive)$
-$typecode.cTypename$
-$else$
-const $typecode.cTypename$&
-$endif$
->>
-
-varDecl(variable, create) ::= <<
-$if(variable)$
-$if(variable.typecode.primitive)$
-$variable.typecode.cTypename$  $variable.name$ = $variable.typecode.initialValue$;
-$else$
-$variable.typecode.cTypename$  $variable.name$;
-$endif$
-$endif$
->>
-
-varDel(variable) ::= <<>>
-
-member_default_init(member) ::= <<
-$if(member.typecode.primitive)$
-m_$member.name$ = $member.typecode.initialValue$;
-$endif$
->>
-
-getTypenameOperation(operation, suffix, varname) ::= <<
-$varname$ = $operation.parent.scopedname$Plugin::$operation.name$$suffix$Plugin::get_typename();
->>
-
-getTypenameInterface(interface, suffix, varname) ::= <<
-$varname$ = $interface.scopedname$$suffix$Plugin::get_typename();
->>
-
-registerTypenameOperationAndCheck(operation, suffix, varname) ::= <<
-!$operation.parent.scopedname$Plugin::$operation.name$$suffix$Plugin::register_type(m_ddsTransport->getRTPSParticipant(), $varname$)
->>
-
-registerTypenameInterfaceAndCheck(interface, suffix, varname) ::= <<
-!$interface.scopedname$$suffix$Plugin::register_type(m_ddsTransport->getRTPSParticipant(), $varname$)
->>
-
-createDataCallback(operation, suffix) ::= <<
-(::transport::dds::Transport::Create_data)$operation.parent.scopedname$Plugin::$operation.name$$suffix$Plugin::create_data
->>
-
-createDataCallbackInterface(interface, suffix) ::= <<
-(::transport::dds::Transport::Create_data)$interface.scopedname$$suffix$Plugin::create_data
->>
-
-copyDataCallback(operation, suffix) ::= <<
-(::transport::dds::Transport::Copy_data)$operation.parent.scopedname$Plugin::$operation.name$$suffix$Plugin::copy_data
->>
-
-copyDataCallbackInterface(interface, suffix) ::= <<
-(::transport::dds::Transport::Copy_data)$interface.scopedname$$suffix$Plugin::copy_data
->>
-
-destroyDataCallback(operation, suffix) ::= <<
-(::transport::dds::Transport::Destroy_data)$operation.parent.scopedname$Plugin::$operation.name$$suffix$Plugin::destroy_data
->>
-
-destroyDataCallbackInterface(interface, suffix) ::= <<
-(::transport::dds::Transport::Destroy_data)$interface.scopedname$$suffix$Plugin::destroy_data
->>
-
-addParamToTopic(param, varname, isRequest) ::= <<
-$if(param)$
-$if(isRequest)$
-$varname$.$param.name$($param.name$);
-$else$
-$if(param.typecode.primitive)$
-$varname$.$param.name$($param.name$);
-$else$
-$varname$.$param.name$(std::move($param.name$));
-$endif$
-$endif$
-$endif$
->>
-
-addExceptionMemberToTopic(exception, member, varname) ::= <<
-$if(member.typecode.primitive)$
-$varname$.unio().$exception.formatedScopedname$_ex().$member.name$(ex.$member.name$());
-$else$
-$varname$.unio().$exception.formatedScopedname$_ex().$member.name$(std::move(ex.$member.name$()));
-$endif$
->>
-
-extractParamFromTopic(param, varname, isReply) ::= <<
-$if(param.typecode.primitive)$
-$param.name$ = $varname$.$param.name$();
-$else$ 
-$param.name$ = std::move($varname$.$param.name$()); //TODO Probar que esto funciona.
-$endif$ 
->>
-
-extractExceptionMemberFromTopic(exception, member, varname) ::= <<
-$if(member.typecode.primitive)$
-ex.$member.name$($varname$.unio().$exception.formatedScopedname$_ex().$member.name$());
-$else$
-ex.$member.name$(std::move($varname$.unio().$exception.formatedScopedname$_ex().$member.name$()));
-$endif$
->>
-
-extractExceptionMemberFromTopicByInterface(exception, member, varname) ::= <<
-$if(member.typecode.primitive)$
-ex.$member.name$($varname$.$exception.formatedScopedname$_ex().$member.name$());
-$else$
-ex.$member.name$(std::move($varname$.$exception.formatedScopedname$_ex().$member.name$()));
-$endif$
->>
-
-copyHeaderInfo(reqvarname, repvarname) ::= <<
-$repvarname$._header().clientId().value_1() = $reqvarname$._header().clientId().value_1();
-$repvarname$._header().clientId().value_2() = $reqvarname$._header().clientId().value_2();
-$repvarname$._header().clientId().value_3() = $reqvarname$._header().clientId().value_3();
-$repvarname$._header().clientId().value_4() = $reqvarname$._header().clientId().value_4();
-$repvarname$._header().requestSequenceNumber() = $reqvarname$._header().requestSequenceNumber();
->>
-
-setHeaderSuccess(varname) ::= <<
-$varname$._header().retCode(OPERATION_SUCCESSFUL);
->>
-
-setHeaderServerInternalException(varname) ::= <<
-$varname$._header().retCode(SERVER_INTERNAL_ERROR);
-$varname$._header().retMsg(std::move(ex.what()));
->>
-
-getHeaderRetCode(varname) ::= <<
-$varname$._header().retCode()
->>
-
-getHeaderRetMsg(varname) ::= <<
-$varname$._header().retMsg()
->>
-
-interfaceUnionDiscriminator(varname) ::= <<
-$varname$.unio()._d()
->>
-
-interfaceUnionOperation(operation, varname) ::= <<
-$varname$.unio().$operation.name$()
->>
-
-replyUnionDiscriminator(varname) ::= <<
-$varname$.unio()._d()
->>
-
-replyUnionDiscriminatorByInterface(operation, varname) ::= <<
-$varname$.unio().$operation.name$()._d()
->>
-
-replyUnionData(varname, dataname) ::= <<
-$varname$.unio().$dataname$()
->>
-
-replyUnionDataByInterface(operation, varname, dataname) ::= <<
-$varname$.unio().$operation.name$().$dataname$()
->>
-
-eprosimaTypesBool() ::= <<true>>
-
-/** NEEDED BY RTI TYPES **/
-charCasting() ::= <<>>
-
-unsetReply() ::= <<>>
-
-reqrepDestruction(type, reqrep, name, void) ::= <<>>
-
-varComplexInitializeT(typeName, name) ::= <<>>
-
-varComplexInitialize(variable) ::= <<>>
-
-varComplexDelT(typeName, name) ::= <<>>
-
-reqrepRefFinalize(type, reqrep, name) ::= <<>>
