--- src/com/eprosima/micrortps/idl/templates/idlTypes.stg
@@ -1,59 +1,0 @@
-// Copyright 2016 Proyectos y Sistemas de Mantenimiento SL (eProsima).
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-group Types;
-
-type_5() ::= <<float>>
-
-type_6() ::= <<double>>
-
-type_13() ::= <<long double>>
-
-type_1() ::= <<short>>
-
-type_2() ::= <<long>>
-
-type_11() ::= <<long long>>
-
-type_3() ::= <<unsigned short>>
-
-type_4() ::= <<unsigned long>>
-
-type_12() ::= <<unsigned long long>>
-
-type_8() ::= <<char>>
-
-type_14() ::= <<wchar>>
-
-type_7() ::= <<boolean>>
-
-type_9() ::= <<octet>>
-
-type_d() ::= <<string>>
-
-type_15() ::= <<wstring>>
-
-type_f_first(prev) ::= <<std::array<$prev$>>
-type_f_second(prev, size) ::= <<, $size$>$prev$>>
-type_f(firs, secon, type) ::= <<$firs$$type$$secon$>>
-
-type_a(name) ::= <<$name$>>
-
-type_b(name) ::= <<$name$>>
-
-type_c(name) ::= <<$name$>>
-
-type_10(name) ::= <<$name$>>
-
-type_e(type, maxsize, empty) ::= <<sequence<$type$$if(maxsize)$, $maxsize$$endif$>$empty$>>
