+++ src/main/java/com/eprosima/micrortps/idl/parser/typecode/StructTypeCode.java
