+++ src/main/java/com/eprosima/micrortps/util/VSConfiguration.java
