+++ src/main/resources/com/eprosima/micrortps/idl/templates/WriteAccessHeader.stg
