+++ src/main/resources/com/eprosima/micrortps/idl/templates/WriteAccessSource.stg
