+++ test/TestType.idl
@@ -1,5 +1,11 @@
 #include "IncludeType.idl"
 
+struct CustomType
+{
+    long                       custom_long_1;
+    long                       custom_long_2;
+};
+
 struct TestType {
     float                      test_float;
     double                     test_double;
@@ -13,7 +19,16 @@ struct TestType {
     boolean                    test_boolean;
     octet                      test_octet;
     string                     test_string;
-    IncludeType                test_include_type;
-    sequence<long>             test_seq_long;
+    CustomType                 test_custom;
+    sequence<long, 8>          test_seq_long;
+    sequence<CustomType, 8>    test_seq_custom;
+    sequence<string, 8>        test_seq_string;
     long                       test_array_long[2];
+    CustomType                 test_array_custom[2];
+    string                     test_array_string[2];
+    long                       test_array_2_long[2][4];
+    CustomType                 test_array_2_custom[2][4];
+    string                     test_array_2_string[2][4];
+    long                       test_array_3_long[2][4][8];
 };
+
