+++ utils/releaseTest/ArrayIDL/ArrayIDL.idl
@@ -1,0 +1,16 @@
+struct Custom
+{
+    long a;
+    long b;
+};
+
+struct S1
+{
+    long aarraybasic[64];
+    Custom aarraycustom[64];
+    string aarraystring[64];
+    long aarraybasic2[64][128];
+    Custom aarraycustom2[64][128];
+    string aarraystring2[64][128];
+    long aarraybasic3[64][128][256];
+};
