+++ utils/releaseTest/IncludeIDL/IncludeIDL1.idl
@@ -2,17 +2,3 @@ struct S1
 {
 	long value;
 };
-
-union UN1 switch(long)
-{
-	case 1:
-		long value;
-	default:
-		boolean abool;
-};
-
-enum MYENUM
-{
-	HOLA,
-	ADIOS
-};
