+++ utils/releaseTest/IncludeIDL/IncludeIDL2.idl
@@ -4,6 +4,4 @@
 struct S2
 {
 	S1 astru;
-	UN1 auni;
-	MYENUM aenum;
 };
