+++ utils/releaseTest/StringIDL/StringIDL.idl
@@ -1,0 +1,5 @@
+struct StringContainer
+{
+    string str1;
+    string<1024> str2;
+};
